Fix type of $timeout parameter of BaseJsonClient::receive
Php::Type::Numeric is an integer, not a float.
See:
- https://github.com/yaroslavche/phptdlib/blob/0.0.10/tdlib.cpp#L34
- https://github.com/CopernicaMarketingSoftware/PHP-CPP/blob/v2.2.0/include/type.h#L25

+++ stubs/TDLib/BaseJsonClient.php
@@ -27,10 +27,10 @@ class BaseJsonClient {
     }
 
     /**
-     * @param float $timeout
+     * @param int $timeout
      * @return string
      */
-    public function receive(float $timeout):string
+    public function receive(int $timeout):string
     {
     }
 }
